feat: add CORS middleware, single list retrieval, and ObjectId validation

+++ day-1-curd/server/src/controller/list.controller.js
@@ -1,6 +1,7 @@
+import mongoose from "mongoose";
 import listModel from "../models/list.model.js";
 
-export let createListController = async () => {
+export let createListController = async (req, res) => {
   try {
     const { taskName, description } = req.body;
     if (!taskName || !description) {
@@ -27,7 +28,7 @@ export let createListController = async () => {
   }
 };
 
-export let getAllListController = async () => {
+export let getAllListController = async (req, res) => {
   try {
     const allList = await listModel.find();
 
@@ -51,9 +52,47 @@ export let getAllListController = async () => {
   }
 };
 
-export let updateList = async () => {
+export let getOneListController = async (req, res) => {
   try {
     const { taskId } = req.params;
+
+    if (!mongoose.Types.ObjectId.isValid(taskId)) {
+      return res.status(400).json({
+        success: false,
+        message: "Invalid Task ID format",
+      });
+    }
+
+    const oneList = await listModel.findById(taskId);
+    if (!oneList) {
+      return res.status(404).json({
+        success: false,
+        message: "List not found",
+      });
+    }
+    return res.status(200).json({
+      success: true,
+      message: "List fetched successfully",
+      data: oneList,
+    });
+  } catch (error) {
+    return res.status(500).json({
+      message: "internal server error",
+    });
+  }
+};
+
+export let updateList = async (req, res) => {
+  try {
+    const { taskId } = req.params;
+
+    if (!mongoose.Types.ObjectId.isValid(taskId)) {
+      return res.status(400).json({
+        success: false,
+        message: "Invalid Task ID format",
+      });
+    }
+
     const { taskName, description, isCompleted } = req.body;
     const updatedList = await listModel.findByIdAndUpdate(
       taskId,
@@ -72,11 +111,19 @@ export let updateList = async () => {
   }
 };
 
-export let deleteList = async () => {
+export let deleteList = async (req, res) => {
   try {
     const { taskId } = req.params;
+
+    if (!mongoose.Types.ObjectId.isValid(taskId)) {
+      return res.status(400).json({
+        success: false,
+        message: "Invalid Task ID format",
+      });
+    }
+
     const deletedList = await listModel.findByIdAndDelete(taskId);
-    if (!deleteList) {
+    if (!deletedList) {
       return res.status(400).json({
         success: false,
         message: "List not found",
